High: LRM: lrmd,clientlib: asynchronous resource delete notification (lf#2439)
This changeset entails a change in resource delete, op cancel,
and all op flush requests.

Resource cannot be deleted if there is an operation running on
the resource. The same obviously for op cancel/flush.

The following library calls may return HA_RSCBUSY in addition to
HA_OK and HA_FAIL:

- delete_rsc (lrm_delete_rsc)
- cancel_op (rsc_cancel_op)
- flush_ops (rsc_flush_ops)

In addition, on delayed resource delete, op->rsc_deleted will be
set to 1 for the operation which was keeping the resource busy.