Medium: stonith: prevent symbol conflict with libcrypto (thanks to lge) (lf#2571, bnc#694243)
In 2005, when some early Heartbeat 2.0.x was due, Debian was at
odds with OpenSSL, so that dependency was removed. But, the
functions which moved from the Heartbeat code to the clplumbing
library weren't renamed. The symbols which conflict with
libcrypto are HMAC and MD5. The problem surfaced with the change
in stonith(8), which introduced logging using cl_log.

The ipmilan stonith plugin links with libOpenIPMI.so which links
with libcrypto.so and then instead of using HMAC from libcrypto
ldd resolves it to the one from libplumb.so. HMAC from libplumb
has a different API from the one in OpenSSL so all kinds of
strange things would happen (mainly looking like a stack
corruption).

With this changeset we pick those few logging related symbols by
hand from libplumb.so. That fixes this particular case, but it
doesn't remove the root cause. I guess that that would be
possible only by renaming symbols in clplumbing, then renaming
them in Heartbeat and Pacemaker and other users, then bumping
the library version. Sounds like quite a bit of effort and the
last clplumbing API change proved to be challenging.