Medium: cl_log: log spamming control
Some conditions may persist over long periods and cause huge
number of similar messages to be logged. This feature allows
limiting number of messages within a timeframe. For instance,
maximum of ten messages within one minute. Once this condition is
detected, appropriate message and a predefined advice are logged.
Further messages of the same kind are suppressed for a while,
say two minutes.