fix syslogmsgfmt logging inconsistency for stderr/stdout
For stderr/stdout, print_logline was called unconditionally.
If syslog message format is enabled (which is the default),
that differs from the log format used for other files.

If we instead use append_log(...) for stderr and stdout as well,
we take the same code path as for any other "file" target.